# backend/internal/platform/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TeamOps API
  version: 1.0.0
  description: |
    REST API for multi-tenant project and task management.

    Access tokens are short-lived JWT bearer tokens. Refresh tokens are opaque,
    single-use credentials rotated by `/api/v1/auth/refresh`. Send JSON request
    bodies with `Content-Type: application/json`; unknown properties and trailing
    JSON values are rejected.

    Authorization is enforced server-side. OWNER has complete tenant control;
    ADMIN manages members, projects, tasks, labels, and audit logs but cannot
    delete the organization; MEMBER reads the tenant and creates/updates tasks
    and comments; VIEWER is read-only. Resource endpoints derive the organization
    from persisted projects/tasks to prevent IDOR.

    List endpoints use bounded offset pagination. `sort=-created_at` means
    descending; a `+` prefix means ascending. The optional `order=asc|desc`
    parameter is used when sort has no prefix. Project and task updates require
    the current `version`; stale versions return HTTP 409.

    Every `/api/v1` operation is subject to the configured IP rate limit and
    can return HTTP 429. JSON-writing operations can also return HTTP 413 or
    415, and unexpected server failures use the common error envelope with
    HTTP 500.
servers:
  - url: http://localhost:8080
    description: Local Docker/API server
...
paths:
...
  /api/v1/organizations/{organizationId}/members/{userId}:
    parameters: [{ $ref: '#/components/parameters/OrganizationId' }, { $ref: '#/components/parameters/UserId' }]
    delete:
      tags: [Organizations]
      operationId: removeMember
      summary: Remove a non-owner member (OWNER or ADMIN)
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Member removed }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /api/v1/organizations/{organizationId}/labels:
    parameters: [{ $ref: '#/components/parameters/OrganizationId' }]
    get:
      tags: [Labels]
      operationId: listLabels
      summary: List organization labels
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Labels, content: { application/json: { schema: { $ref: '#/components/schemas/LabelListEnvelope' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
    post:
      tags: [Labels]
      operationId: createLabel
      summary: Create an organization label (OWNER or ADMIN)
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/LabelCreateRequest' } } } }
      responses:
        '201': { description: Label created, content: { application/json: { schema: { $ref: '#/components/schemas/LabelEnvelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/BodyTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT, description: 'Short-lived access token returned by register, login, or refresh.' }
  parameters:
...
    OrganizationId: { name: organizationId, in: path, required: true, schema: { type: string, format: uuid } }
...
    UserId: { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
  responses:
    ValidationError: { description: 'Malformed JSON, unknown field, invalid identifier, invalid filter, or business validation failure', content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' }, example: { error: { code: validation_error, message: request validation failed, requestId: 8c69acaf-fda2-42bd-95c2-59a7ed3153d4 } } } } }
    Unauthorized: { description: 'Missing, malformed, expired, or invalid authentication credential', content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' }, example: { error: { code: unauthorized, message: authentication required, requestId: 8c69acaf-fda2-42bd-95c2-59a7ed3153d4 } } } } }
...
    Forbidden: { description: Authenticated user lacks the required tenant role or resource access, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' }, example: { error: { code: forbidden, message: insufficient permissions, requestId: 8c69acaf-fda2-42bd-95c2-59a7ed3153d4 } } } } }
    NotFound: { description: Resource does not exist or is intentionally concealed from a non-member, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Conflict: { description: 'Optimistic-lock version, uniqueness, immutable-owner, or state-transition conflict', content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' }, example: { error: { code: version_conflict, message: resource was modified by another request, requestId: 8c69acaf-fda2-42bd-95c2-59a7ed3153d4 } } } } }
    BodyTooLarge: { description: Request body exceeds configured limit, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    UnsupportedMediaType: { description: JSON endpoint requires Content-Type application/json, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    RateLimited: { description: IP-based or login abuse limit exceeded, headers: { Retry-After: { schema: { type: integer }, description: Seconds until the fixed window resets }, X-RateLimit-Limit: { schema: { type: integer } }, X-RateLimit-Remaining: { schema: { type: integer } } }, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
  schemas:
    Error: { type: object, additionalProperties: false, required: [code, message], properties: { code: { type: string }, message: { type: string }, requestId: { type: string, maxLength: 128 } } }
    ErrorEnvelope: { type: object, additionalProperties: false, required: [error], properties: { error: { $ref: '#/components/schemas/Error' } } }
...
    LabelCreateRequest: { type: object, additionalProperties: false, required: [name, color], properties: { name: { type: string, minLength: 1, maxLength: 50 }, color: { type: string, pattern: '^#[0-9A-Fa-f]{6}$', example: '#7C3AED' } } }
...
    Label: { type: object, additionalProperties: false, required: [id, organizationId, name, color], properties: { id: { type: string, format: uuid }, organizationId: { type: string, format: uuid }, name: { type: string }, color: { type: string } } }
...
    LabelEnvelope: { type: object, required: [data], properties: { data: { $ref: '#/components/schemas/Label' } } }
    LabelListEnvelope: { type: object, required: [data], properties: { data: { type: array, items: { $ref: '#/components/schemas/Label' } } } }
...